Move tools/verify/ into tool_util/verify/
Ref: https://github.com/galaxyproject/galaxy/issues/8073
