+++ lib/galaxy/tool_util/verify/asserts/archive.py
