+++ lib/galaxy/tool_util/verify/asserts/text.py
